feat: Add discount_price and on_offer fields to Product model.
Add those fields to the admin view. Add product_price_paid to the
OrderLineItem model to store the actual price paid by the user.
Within the OrderLineItem model, if the product is on offer and has
a discounted price, use that to calculate the lineitem_total. Change
the cart_contexts function to check whether an item is on_offer and
has a discounted price, and if so, use that value to augment the
total. Use 'if' 'else' statements on the products.html,
product_detail.html, cart.html, checkout.html and checkout_success.html
pages to ascertain whether the product is on offer and display the
correct price/total.