feat: Set SECRET_KEY and DEBUG variables.
In settings.py, check environ for the SECRET_KEY variable or set to a blank string.
Set DEBUG to be True only if there's a variable called 'development' in the environment.
This ensures that the SECRET_KEY will not be visible in version control.
This also ensures that DEBUG will be False on Heroku.