feat: Add AWS config variables.
In settings.py, check to see whether 'USE_AWS' exists in the environment.
If so, use the AWS and STATICFILES variables.
These direct Django to get static and media files from the S3 Bucket
and to upload static files to the Bucket when 'collectstatic' is run
during the build process.