docs: Deployment write-up of S3 bucket media files and the Stripe webhook.
Document the steps taken to populate the S3 bucket media folder with
product images, blog images and other banner images etc.
Document the process of linking Stripe to Heroku.
Add the Stripe 'public' and 'secret' keys as config variables in Heroku.
Create a new Webhook endpoint in Stripe that points to the Heroku app url.
Add the STRIPE_WH_SECRET to Heroku as a config variable.