refactor: Combine if statements when checking for discount code.
Within contexts.py combine the try, except and if statements
into one if statement that checks whether
the discount_code_object exists and
whether the user_discount_code exists and
whether that user_discount_code is active.
If these three conditions are true, the discount is calculated.
Otherwise the discount is not applied and the discount variable
is deleted from the session cookie.