feat: Redirect users to the previous when editing reviews.
Add a hidden input on the edit_review form to send the
request.META.HTTP_REFERER i.e. the previous page url
to the view.
Then, within the view, check whether the substring 'profile'
is to be found within the url.
If so, then the user accessed the edit_review page from their
profile rather than the product page.
Redirect them accordingly.