feat: Redirect to Profile or Blog from Edit BlogComment form.
Add the referring url to the Edit BlogComment form.
Within the edit_comment view, save this value to the
previous_page_url variable.
Upon successful submission of the form,
check for the existence of the substring profile
within the url. If it exists, redirect to the
user's profile page. Otherwise, they accessed this
page from the blog_post page, so redirect them there.