feat: Add link to Account page from password_change.html.
If the user is logged in, the My Account button is rendered.
This allows users to navigate to their profile page more
easily, and if they successfully change their password they
can navigate back to the profile page using the link as
successful submission of the form redirects to the same page.