fix: Combine tag Q object with product Q objects.
Check to see whether the query entered in the search bar
exactly matches a tag.  If so, return the list of tagged_products.
Then using the OR operator, combine a Q object which searches by id__in
with the other Q objects that search for the search term in the
product ingredients, information, category and brand fields.
If a search term exactly matches a tag it now no longer returns
just those tagged products.