feat: Add image carousel when product has more than one image.
Add Bootstrap carousel to product_detail page.
In the product_detail view, check to see whether
any productimage objects exist in the database
for this product.
Pass the images list to the template.
Add 'if' 'elif' 'else' statements to check whether
the images list is greater than or equal to 1.
If so, render the carousel, with the product main image
as the first image.
If not, check whether the product has a main_image.
Render it if it exists, otherwise, render the
no-image-placeholder.
This allows users to easily see, and navigate through
all the images associated with a product.