feat: Add unittests for home app.
Add tests folder with unittests for
testing whether the views in the home app
return a 200 response and fetch the correct template.
Add a test to check whether the NewsletterSubscription
model is created correctly and returns the email address
Add tests to check whether the NewsletterSubscriptionForm
is valid when data is input and invalid when the email address
is not inserted.
Add tests to check whether the ContactForm is valid when all
three fields are filled and invalid in the case of a missing
name or a missing email address or if the message is left blank.