fix: Change and Log Save Info Checkbox.
Use getElementById to target Save Info checkbox.
Log variable in webhook_handler and checkout views.
Change if statements to check whether the
variable is True before saving the address info
to the user's profile. Now if a user does not
check the box, their profile information will not
be updated.